+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/codestyle/UnnecessaryFullyQualifiedNameRule.java
@@ -5,6 +5,7 @@
 package net.sourceforge.pmd.lang.java.rule.codestyle;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map.Entry;
@@ -189,8 +190,9 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
         } else {
             ASTImportDeclaration firstMatch = findFirstMatch(matches);
 
-            // Could this done to avoid a conflict?
-            if (!isAvoidingConflict(node, name, firstMatch)) {
+            if (!isReferencingInnerNonStaticClass(name, firstMatch)
+                    && !isAvoidingConflict(node, name, firstMatch)) {
+
                 String importStr = firstMatch.getImportedName() + (firstMatch.isImportOnDemand() ? ".*" : "");
                 String type = firstMatch.isStatic() ? "static " : "";
 
@@ -290,6 +292,24 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
         } else if (isJavaLang) {
             // only java.lang.* is implicitly imported, but not e.g. java.lang.reflection.*
             return StringUtils.countMatches(name, '.') == 2;
+        }
+        return false;
+    }
+
+    private boolean isReferencingInnerNonStaticClass(final String name, final ASTImportDeclaration firstMatch) {
+        if (firstMatch.isImportOnDemand() && firstMatch.isStatic() && firstMatch.getType() != null) {
+            String[] nameParts = name.split("\\.");
+            String[] importParts = firstMatch.getImportedName().split("\\.");
+
+            if (nameParts.length == 2 && importParts[importParts.length - 1].equals(nameParts[0])) {
+                Class<?>[] declaredClasses = firstMatch.getType().getDeclaredClasses();
+                for (Class<?> innerClass : declaredClasses) {
+                    if (nameParts[1].equals(innerClass.getSimpleName()) && (innerClass.getModifiers() & Modifier.STATIC) != Modifier.STATIC) {
+                        // the referenced inner class is not static, therefore the static import on demand doesn't match
+                        return true;
+                    }
+                }
+            }
         }
         return false;
     }